+++ NEWS.md
@@ -5,6 +5,8 @@
 ### New features
 - Add support for toggling whether to search in hidden files (`M-s h` in the
   search prompt, or `urgrep-search-hidden-files` globally)
+- Allow setting the search tool to use on the fly when reading the query
+  (`M-s t` in the search prompt)
 
 ### Breaking changes
 - `urgrep-run-command` now takes `directory` and `tool` as optional keys